requirements.txt
discord.py>=2.3.0
openai>=1.0.0
langdetect
python-dotenv
textblob
requests>=2.28.0
aiohttp>=3.8.0